+++ lib/mhu_flutter_camera.dart
@@ -1,3 +1,3 @@
 export 'src/camera.dart';
-export 'src/generated/mhu_flutter_camera.pb.dart';
-export 'src/generated/mhu_flutter_camera.pbmeta.dart';
+export 'proto.dart';
+export 'src/model.dart';