+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+- improved handling of camera permission denied scenario
+
+## 1.0.1
+
+- setting up automated publishing
+
 ## 1.0.0
 
 - Initial version.